# Treat map keys matching the same struct field as duplicates.